feat(commit): allow configuring git diff context lines
Introduces a new `--contextLines` (alias `-cl`) command-line option for
the `commit` command. This allows users to specify the number of context
lines for the git diff, which is used to generate the commit message.

The default is set to 3 to align with git's standard behavior,
replacing the previous hardcoded value of 25. An alias `-c` has also
been added for the `--context` option.