refactor(rules)!: remove CLI argument for rule file selection
The optional `target` argument for the `rules` command was previously
overloaded, serving to specify either the rule file or the output file.
This created ambiguity for the user and complexity in the implementation.

This change refactors the command to have a clearer, single-purpose
interface. The command now always prompts the user to select a rule
file from the available options. The optional `target` argument is now
used exclusively to define the output file.

BREAKING CHANGE: The `rules` command no longer accepts a positional
argument to specify the rule file.

Previously, `gitai rules <path>` could be used to select a rule
file directly. This functionality has been removed. The command now
always interactively prompts for rule file selection. The optional
positional argument is now solely for specifying the output target file.